+++ husky_components/column-navigation/main.js
@@ -47,8 +47,9 @@
  * @params {Boolean} [options.tooltipTranslations.unpublished] translation-keys for unpublished
  * @params {Boolean} [options.tooltipTranslations.internalLink] translation-keys for internal-link
  * @params {Boolean} [options.tooltipTranslations.externalLink] translation-keys for external-link
+ * @params {Callback} [options.actionCallback] callback which will be called on action
  */
-define([], function() {
+define(function() {
 
     'use strict';
 
@@ -87,7 +88,8 @@ define([], function() {
                 unpublished: 'public.unpublished',
                 internalLink: 'public.internal-link',
                 externalLink: 'public.external-link'
-            }
+            },
+            actionCallback: null
         },
 
         constants = {
@@ -153,7 +155,7 @@ define([], function() {
             item: [
                 '<li data-id="<%= id %>" class="' + constants.columnItemClass + '">',
                 '    <span class="' + constants.iconsLeftClass + '"></span>',
-                '    <span title="<%= title %>" class="' + constants.itemTextClass + '"><%= title%></span>',
+                '    <span title="<%= title %>" class="' + constants.itemTextClass + '"><%= title %></span>',
                 '    <span class="' + constants.iconsRightClass + '"></span>',
                 '</li>'
             ].join(''),
@@ -293,6 +295,15 @@ define([], function() {
          */
         RESIZE = function() {
             return createEventName.call(this, 'resize');
+        },
+
+        /**
+         * @event husky.column-navigation.set-options
+         * @description the element will be rerendered with given options
+         * @param {{url, selected}} options
+         */
+        SET_OPTIONS = function() {
+            return createEventName.call(this, 'set-options');
         },
 
         createContext = function(column) {
@@ -678,7 +689,7 @@ define([], function() {
          */
         renderItem: function(data) {
             var $item = this.sandbox.dom.createElement(this.sandbox.util.template(templates.item)({
-                    title: this.sandbox.util.escapeHtml(data[this.options.titleName]),
+                    title: this.sandbox.translate(this.sandbox.util.escapeHtml(data[this.options.titleName])),
                     id: data[this.options.idName]
                 })),
                 disabled = (this.options.disableIds.indexOf(data[this.options.idName]) !== -1);
@@ -930,6 +941,22 @@ define([], function() {
             }
         },
 
+        unbindDOMEvents: function() {
+            this.$el.off();
+            if (!!this.dom.$add) {
+                this.dom.$add.off();
+            }
+            if (!!this.dom.$container) {
+                this.dom.$container.off();
+            }
+            if (!!this.dom.$wrapper) {
+                this.dom.$wrapper.off();
+            }
+            if (!!this.dom.$ok) {
+                this.dom.$ok.off();
+            }
+        },
+
         /**
          * Handles the key-down event of a order-input
          * @param event
@@ -940,7 +967,7 @@ define([], function() {
             }
             if (event.keyCode === 27) { // cancel on esc
                 var column = this.sandbox.dom.attr(this.sandbox.dom.parents(
-                        event.currentTarget, '.' + constants.columnClass), 'data-column'),
+                    event.currentTarget, '.' + constants.columnClass), 'data-column'),
                     item = this.sandbox.dom.attr(this.sandbox.dom.parents(
                         event.currentTarget, '.' + constants.columnItemClass), 'data-id');
                 this.resetOrderInput(column, item);
@@ -1106,10 +1133,21 @@ define([], function() {
         },
 
         bindCustomEvents: function() {
+            if (!!this.customEvents) {
+                return;
+            }
+
             this.sandbox.on(BREADCRUMB.call(this), this.getBreadCrumb.bind(this));
             this.sandbox.on(UNMARK.call(this), this.unmark.bind(this));
             this.sandbox.on(HIGHLIGHT.call(this), this.highlight.bind(this));
             this.sandbox.on(ORDER.call(this), this.startOrderModeItem.bind(this));
+            this.sandbox.on(SET_OPTIONS.call(this), function(options) {
+                this.unbindDOMEvents();
+                this.$el.children().remove();
+                this.options = this.sandbox.util.extend(true, {}, this.options, options);
+
+                this.initialize();
+            }.bind(this));
 
             this.sandbox.on('husky.dropdown.' + this.options.instanceName + '.settings.dropdown.item.click', this.dropdownItemClicked.bind(this));
 
@@ -1119,6 +1157,8 @@ define([], function() {
                     this.setOverflowClass();
                 }.bind(this));
             }
+
+            this.customEvents = true;
         },
 
         /**
@@ -1410,7 +1450,7 @@ define([], function() {
                 }
 
                 // scroll for add column
-                if (!selectedItem.hasSub) {
+                if (!selectedItem[this.options.hasSubName]) {
                     this.insertAddColumn(selectedItem, column);
                     this.scrollIfNeeded(column);
                     this.setOverflowClass();
@@ -1498,7 +1538,11 @@ define([], function() {
             }
 
             this.sandbox.dom.stopPropagation(event);
-            this.sandbox.emit(ACTION.call(this), item);
+            if (!!this.options.actionCallback) {
+                this.options.actionCallback(item)
+            } else {
+                this.sandbox.emit(ACTION.call(this), item);
+            }
         }
     };
 });