+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG for husky
 
 * 0.17.0 (2015-11-18)
+    * FEATURE     #580 Added set options event for column-navigation
     * ENHANCEMENT #580 Added events to data navigation for selecting items
     * BUGFIX      #576 Fixed update of data navigation