+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG for husky
 
-* dev-develop
+* 0.17.0 (2015-12-01)
     * ENHANCEMENT #584 Implemented reset event for toolbar button
     * FEATURE     #583 Added new input skin for locked-inputs
     * BUGFIX      #579 Fixed datagrid error when window resize is emitted before initialization